Move tag position management into state manager. Might have side-effect in LyricEditor but let it pass now.
