+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricEditorStateManager.cs
@@ -1,13 +1,18 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Linq;
 using osu.Framework.Bindables;
+using osu.Framework.Extensions.IEnumerableExtensions;
 using osu.Game.Rulesets.Karaoke.Objects;
+using osu.Game.Screens.Edit;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 {
     public class LyricEditorStateManager
     {
+        private EditorBeatmap beatmap { get; set; }
+
         public Bindable<Lyric> BindableSplitLyric { get; } = new Bindable<Lyric>();
         public Bindable<int> BindableSplitPosition { get; } = new Bindable<int>();
 
@@ -18,6 +23,15 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
         public Mode Mode => BindableMode.Value;
 
         public LyricFastEditMode FastEditMode => BindableFastEditMode.Value;
+
+        public Bindable<TimeTag> BindableCursorPosition { get; } = new Bindable<TimeTag>();
+
+        public LyricEditorStateManager(EditorBeatmap beatmap)
+        {
+            this.beatmap = beatmap;
+        }
+
+        #region Cursor position by time-tag
 
         public void UpdateSplitCursorPosition(Lyric lyric, int index)
         {
@@ -30,6 +44,117 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
             BindableSplitLyric.Value = null;
         }
 
+        #endregion
+
+        #region Cursor position by lyric
+
+        public bool MoveCursor(CursorAction action)
+        {
+            var currentTimeTag = BindableCursorPosition.Value;
+
+            TimeTag nextTimeTag = null;
+
+            switch (action)
+            {
+                case CursorAction.MoveUp:
+                    nextTimeTag = getPreviousLyricTimeTag(currentTimeTag);
+                    break;
+
+                case CursorAction.MoveDown:
+                    nextTimeTag = getNextLyricTimeTag(currentTimeTag);
+                    break;
+
+                case CursorAction.MoveLeft:
+                    nextTimeTag = getPreviousTimeTag(currentTimeTag);
+                    break;
+
+                case CursorAction.MoveRight:
+                    nextTimeTag = getNextTimeTag(currentTimeTag);
+                    break;
+
+                case CursorAction.First:
+                    nextTimeTag = getFirstTimeTag(currentTimeTag);
+                    break;
+
+                case CursorAction.Last:
+                    nextTimeTag = getLastTimeTag(currentTimeTag);
+                    break;
+            }
+
+            if (nextTimeTag == null)
+                return false;
+
+            moveCursorTo(nextTimeTag);
+            return true;
+        }
+
+        public bool MoveCursorToTargetPosition(TimeTag timeTag)
+        {
+            if (timeTagInLyric(timeTag) == null)
+                return false;
+
+            moveCursorTo(timeTag);
+            return true;
+        }
+
+        private Lyric timeTagInLyric(TimeTag timeTag)
+        {
+            if (timeTag == null)
+                return null;
+
+            return beatmap.HitObjects.OfType<Lyric>().FirstOrDefault(x => x.TimeTags?.Contains(timeTag) ?? false);
+        }
+
+        private TimeTag getPreviousLyricTimeTag(TimeTag timeTag)
+        {
+            var lyrics = beatmap.HitObjects.OfType<Lyric>().ToList();
+            var currentLyric = timeTagInLyric(timeTag);
+            return lyrics.GetPrevious(currentLyric)?.TimeTags?.FirstOrDefault(x => x.Index >= timeTag.Index);
+        }
+
+        private TimeTag getNextLyricTimeTag(TimeTag timeTag)
+        {
+            var lyrics = beatmap.HitObjects.OfType<Lyric>().ToList();
+            var currentLyric = timeTagInLyric(timeTag);
+            return lyrics.GetNext(currentLyric)?.TimeTags?.FirstOrDefault(x => x.Index >= timeTag.Index);
+        }
+
+        private TimeTag getPreviousTimeTag(TimeTag timeTag)
+        {
+            var timeTags = beatmap.HitObjects.OfType<Lyric>().SelectMany(x => x.TimeTags).ToArray();
+            return timeTags.GetPrevious(timeTag);
+        }
+
+        private TimeTag getNextTimeTag(TimeTag timeTag)
+        {
+            var timeTags = beatmap.HitObjects.OfType<Lyric>().SelectMany(x => x.TimeTags).ToArray();
+            return timeTags.GetNext(timeTag);
+        }
+
+        private TimeTag getFirstTimeTag(TimeTag timeTag)
+        {
+            var timeTags = beatmap.HitObjects.OfType<Lyric>().SelectMany(x => x.TimeTags).ToArray();
+            return timeTags.FirstOrDefault();
+        }
+
+        private TimeTag getLastTimeTag(TimeTag timeTag)
+        {
+            var timeTags = beatmap.HitObjects.OfType<Lyric>().SelectMany(x => x.TimeTags).ToArray();
+            return timeTags.LastOrDefault();
+        }
+
+        private void moveCursorTo(TimeTag timeTag)
+        {
+            if (timeTag == null)
+                return;
+
+            BindableCursorPosition.Value = timeTag;
+        }
+
+        #endregion
+
+        #region Mode
+
         public void SetMode(Mode mode)
         {
             BindableMode.Value = mode;
@@ -39,6 +164,8 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
         {
             BindableFastEditMode.Value = fastEditMode;
         }
+
+        #endregion
     }
 
     public enum Mode
